requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.9.0
Flask-Login==0.2.9
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.3
Tempita==0.5.3dev
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
blinker==1.3
decorator==3.4.0
distribute==0.6.34
gunicorn==18.0
itsdangerous==0.23
pbr==0.6
sqlalchemy-migrate==0.8.2
wsgiref==0.1.2
